+++ src/check_jsonschema/instance_loader.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import pathlib
 import typing as t
 
 from .parsers import ParseError, ParserSet
@@ -10,11 +11,11 @@ class InstanceLoader:
     def __init__(
         self,
         filenames: t.Sequence[str],
-        default_filetype: str | None = None,
+        default_filetype: str = "json",
         data_transform: Transform | None = None,
     ) -> None:
         self._filenames = filenames
-        self._default_ft = default_filetype.lower() if default_filetype else None
+        self._default_filetype = default_filetype
         self._data_transform = (
             data_transform if data_transform is not None else Transform()
         )
@@ -23,12 +24,13 @@ class InstanceLoader:
             modify_yaml_implementation=self._data_transform.modify_yaml_implementation
         )
 
-    def iter_files(self) -> t.Iterator[tuple[str, ParseError | t.Any]]:
+    def iter_files(self) -> t.Iterator[tuple[pathlib.Path, ParseError | t.Any]]:
         for fn in self._filenames:
+            path = pathlib.Path(fn)
             try:
-                data: t.Any = self._parsers.parse_file(fn, self._default_ft)
+                data: t.Any = self._parsers.parse_file(path, self._default_filetype)
             except ParseError as err:
                 data = err
             else:
                 data = self._data_transform(data)
-            yield (fn, data)
+            yield (path, data)